+++ test/AresTreasury.t.sol
@@ -5,7 +5,6 @@ import "forge-std/Test.sol";
 import "../src/core/AresTreasury.sol";
 
 contract AresTreasuryTest is Test {
-
     AresTreasury treasury;
     address governor;
     uint256 governorPk;
@@ -15,13 +14,12 @@ contract AresTreasuryTest is Test {
     uint256 attackerPk;
 
     function setUp() public {
-
         // deterministically create addresses
         (governor, governorPk) = makeAddrAndKey("governor");
         user1 = makeAddr("user1");
         user2 = makeAddr("user2");
         (attacker, attackerPk) = makeAddrAndKey("attacker");
-        
+
         vm.prank(governor);
         treasury = new AresTreasury();
 
@@ -30,7 +28,6 @@ contract AresTreasuryTest is Test {
     }
 
     receive() external payable {}
-
 
     function testProposalLifecycle() public {
         vm.prank(governor);
@@ -60,7 +57,7 @@ contract AresTreasuryTest is Test {
         vm.prank(governor);
         treasury.execute(id, sig);
 
-        (, , , , , , bool executed, ) = treasury.proposals(id);
+        (,,,,,, bool executed,) = treasury.proposals(id);
         assertTrue(executed);
     }
 
@@ -71,7 +68,7 @@ contract AresTreasuryTest is Test {
         vm.prank(governor);
         treasury.cancel(id);
 
-        (, , , , , , , bool cancelled) = treasury.proposals(id);
+        (,,,,,,, bool cancelled) = treasury.proposals(id);
         assertTrue(cancelled);
     }
 
@@ -99,8 +96,8 @@ contract AresTreasuryTest is Test {
 
     function testRewardClaim() public {
         // setup merkle root with a real 2-leaf tree
-        (bytes32 root, bytes32[] memory proof1, ) = _generateMerkleTree(user1, 1 ether, user2, 2 ether);
-        
+        (bytes32 root, bytes32[] memory proof1,) = _generateMerkleTree(user1, 1 ether, user2, 2 ether);
+
         vm.prank(governor);
         treasury.updateMerkleRoot(root);
 
@@ -111,8 +108,9 @@ contract AresTreasuryTest is Test {
     }
 
     function testMultipleSuccessfulClaims() public {
-        (bytes32 root, bytes32[] memory proof1, bytes32[] memory proof2) = _generateMerkleTree(user1, 1 ether, user2, 2 ether);
-        
+        (bytes32 root, bytes32[] memory proof1, bytes32[] memory proof2) =
+            _generateMerkleTree(user1, 1 ether, user2, 2 ether);
+
         vm.prank(governor);
         treasury.updateMerkleRoot(root);
 
@@ -132,8 +130,8 @@ contract AresTreasuryTest is Test {
     }
 
     function testDoubleClaimReverts() public {
-        (bytes32 root, bytes32[] memory proof1, ) = _generateMerkleTree(user1, 1 ether, user2, 2 ether);
-        
+        (bytes32 root, bytes32[] memory proof1,) = _generateMerkleTree(user1, 1 ether, user2, 2 ether);
+
         vm.prank(governor);
         treasury.updateMerkleRoot(root);
 
@@ -184,10 +182,11 @@ contract AresTreasuryTest is Test {
         vm.expectRevert();
         treasury.execute(id, sig);
     }
+
     function testProposeMaxLimit() public {
         vm.prank(governor);
         bytes32 id = treasury.propose(address(this), 100 ether, "");
-        (, , uint256 value, , , , , ) = treasury.proposals(id);
+        (,, uint256 value,,,,,) = treasury.proposals(id);
         assertEq(value, 100 ether);
     }
 
@@ -196,7 +195,6 @@ contract AresTreasuryTest is Test {
         vm.expectRevert();
         treasury.propose(address(this), 101 ether, "");
     }
-
 
     function testQueueCanceledProposalReverts() public {
         vm.prank(governor);
@@ -244,8 +242,8 @@ contract AresTreasuryTest is Test {
         vm.prank(governor);
         treasury.execute(id2, abi.encodePacked(r2, s2, v2));
 
-        (, , , , , , bool executed1, ) = treasury.proposals(id1);
-        (, , , , , , bool executed2, ) = treasury.proposals(id2);
+        (,,,,,, bool executed1,) = treasury.proposals(id1);
+        (,,,,,, bool executed2,) = treasury.proposals(id2);
         assertTrue(executed1);
         assertTrue(executed2);
     }
@@ -256,7 +254,7 @@ contract AresTreasuryTest is Test {
 
     //     vm.prank(governor);
     //     treasury.queue(id);
-        
+
     //     vm.warp(block.timestamp + 2 days);
 
     //     bytes32 digest = _getDigest(id, 0);
@@ -272,19 +270,12 @@ contract AresTreasuryTest is Test {
     //     treasury.execute(id, sig);
     // }
 
-
     function _getDigest(bytes32 id, uint256 nonce) internal view returns (bytes32) {
         return keccak256(
             abi.encodePacked(
                 "\x19\x01",
                 treasury.DOMAIN_SEPARATOR(),
-                keccak256(
-                    abi.encode(
-                        keccak256("Execute(bytes32 proposal,uint256 nonce)"),
-                        id,
-                        nonce
-                    )
-                )
+                keccak256(abi.encode(keccak256("Execute(bytes32 proposal,uint256 nonce)"), id, nonce))
             )
         );
     }
@@ -292,7 +283,7 @@ contract AresTreasuryTest is Test {
     function testQueueAlreadyQueuedReverts() public {
         vm.prank(governor);
         bytes32 id = treasury.propose(address(this), 1 ether, "");
-        
+
         vm.prank(governor);
         treasury.queue(id);
 
@@ -310,8 +301,8 @@ contract AresTreasuryTest is Test {
         vm.prank(governor);
         treasury.updateMerkleRoot(root);
 
-        bytes32[] memory proof; 
-        
+        bytes32[] memory proof;
+
         vm.expectRevert(AresTreasury.transaction_failed.selector);
         claimer.tryClaim();
     }
@@ -324,7 +315,7 @@ contract AresTreasuryTest is Test {
         vm.warp(block.timestamp + 2 days);
 
         bytes memory sig = new bytes(64); // Invalid length
-        
+
         vm.expectRevert(abi.encodeWithSignature("InvalidSignature()"));
         treasury.execute(id, sig);
     }
@@ -337,8 +328,8 @@ contract AresTreasuryTest is Test {
         vm.warp(block.timestamp + 2 days);
 
         bytes32 digest = _getDigest(id, 0);
-        (uint8 v, bytes32 r, ) = vm.sign(governorPk, digest);
-        
+        (uint8 v, bytes32 r,) = vm.sign(governorPk, digest);
+
         // Malleable S signature
         uint256 invalidS = 0x8000000000000000000000000000000000000000000000000000000000000000;
         bytes memory sig = abi.encodePacked(r, bytes32(invalidS), v);
@@ -384,29 +375,28 @@ contract AresTreasuryTest is Test {
 
     function testClaimReentrancyFails() public {
         ReentrantClaimer claimer = new ReentrantClaimer(treasury);
-        (bytes32 root, bytes32[] memory proof1, ) = _generateMerkleTree(address(claimer), 1 ether, user2, 2 ether);
-        
+        (bytes32 root, bytes32[] memory proof1,) = _generateMerkleTree(address(claimer), 1 ether, user2, 2 ether);
+
         vm.prank(governor);
         treasury.updateMerkleRoot(root);
-        
+
         vm.expectRevert(AresTreasury.transaction_failed.selector);
         claimer.tryClaim(proof1);
     }
 
-    function _generateMerkleTree(
-        address account1, uint256 amount1, 
-        address account2, uint256 amount2
-    ) internal pure returns (bytes32 root, bytes32[] memory proof1, bytes32[] memory proof2) {
+    function _generateMerkleTree(address account1, uint256 amount1, address account2, uint256 amount2)
+        internal
+        pure
+        returns (bytes32 root, bytes32[] memory proof1, bytes32[] memory proof2)
+    {
         bytes32 leaf1 = keccak256(abi.encode(account1, amount1));
         bytes32 leaf2 = keccak256(abi.encode(account2, amount2));
-        
-        root = leaf1 < leaf2 
-            ? keccak256(abi.encodePacked(leaf1, leaf2)) 
-            : keccak256(abi.encodePacked(leaf2, leaf1));
-            
+
+        root = leaf1 < leaf2 ? keccak256(abi.encodePacked(leaf1, leaf2)) : keccak256(abi.encodePacked(leaf2, leaf1));
+
         proof1 = new bytes32[](1);
         proof1[0] = leaf2;
-        
+
         proof2 = new bytes32[](1);
         proof2[0] = leaf1;
     }
@@ -457,7 +447,7 @@ contract ReentrantClaimer {
     }
 
     function tryClaim(bytes32[] memory proof) external {
-        for (uint i = 0; i < proof.length; i++) {
+        for (uint256 i = 0; i < proof.length; i++) {
             savedProof.push(proof[i]);
         }
         treasury.claim(1 ether, proof);
@@ -465,7 +455,7 @@ contract ReentrantClaimer {
 
     receive() external payable {
         bytes32[] memory memProof = new bytes32[](savedProof.length);
-        for (uint i = 0; i < savedProof.length; i++) {
+        for (uint256 i = 0; i < savedProof.length; i++) {
             memProof[i] = savedProof[i];
         }
         treasury.claim(1 ether, memProof);
